Remove AppRater
Since this is a fork of the app, not the app, and it's not distributed through the Play Store, AppRater should be removed. This change should make it not occur (haven't tested it), and util\AppRater.java is probably safe to delete entirely after this change.

PS - Issues is not enabled on this repo. :)